finished parsing employers and employees and preferences

+++ src/main/java/edu/ucmo/spring_example/model/BestMatchEmployee.java
@@ -1,11 +1,12 @@
 package edu.ucmo.spring_example.model;
 
 
+import java.util.HashMap;
 
 public class BestMatchEmployee {
     public int employeeId;
-    public float[] rankings;
-    public BestMatchEmployee(int id, float[] inputRankings) {
+    public int[] rankings;
+    public BestMatchEmployee(int id, int[] inputRankings) {
         employeeId = id;
         rankings = inputRankings;
     }
